Changed : tiny bit of cleaning Changed : SYNC has an RxToggle getter Changed : moved TIME test to other node_test file
