+++ pkg/time/time_test.go
@@ -5,11 +5,8 @@ import (
 	"testing"
 	"time"
 
-	// import assert
 	"github.com/stretchr/testify/assert"
 )
-
-var test = time.Time{}
 
 func TestSetInternalTime(t *testing.T) {
 	now := time.Now()
